[simulation]Export d'odb compatible avec la prepa (ecrire des entiers quand on veut lire des entiers). Merged revision(s) 30183 from branches/sword-scipio-stab

+++ src/libraries/simulation_kernel/Entities/Agents/Units/Dotations/PHY_DotationGroup.cpp
@@ -140,7 +140,7 @@ void PHY_DotationGroup::WriteODB( xml::xostream& xos ) const
     {
         xos << xml::start( "resource" )
                 << xml::attribute( "name", it->first->GetName() )
-                << xml::attribute( "quantity", it->second->GetValue() )
+                << xml::attribute( "quantity", static_cast< int >( it->second->GetValue() ) )
                 << xml::attribute( "logistic-threshold", it->second->GetSupplyThresholdPercentage() );
         xos << xml::end; // dotation
     }
